feat: `VAllocation` form, Separate business partner component.
